remove PaymentRepository.kt
