--- wallet_domain/src/commonMain/kotlin/net/thechance/mena/wallet/domain/repository/PaymentRepository.kt
@@ -1,9 +1,0 @@
-package net.thechance.mena.wallet.domain.repository
-
-import kotlin.uuid.ExperimentalUuidApi
-import kotlin.uuid.Uuid
-
-@OptIn(ExperimentalUuidApi::class)
-interface PaymentRepository {
-    suspend fun submitTransaction(transactionId: Uuid)
-}
