+++ wallet_domain/src/commonMain/kotlin/net/thechance/mena/wallet/domain/repository/TransactionRepository.kt
@@ -22,4 +22,5 @@ interface TransactionRepository {
         amount: Double
     ): Uuid
     suspend fun getTransactionReceiver(transactionId: Uuid): TransactionReceiver
+    suspend fun submitTransaction(transactionId: Uuid)
 }
