+++ wallet_presentation/src/commonMain/kotlin/net/thechance/mena/wallet/presentation/screen/payment_result/PaymentResultViewModel.kt
@@ -5,7 +5,7 @@ package net.thechance.mena.wallet.presentation.screen.payment_result
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.IO
-import net.thechance.mena.wallet.domain.repository.PaymentRepository
+import net.thechance.mena.wallet.domain.repository.TransactionRepository
 import net.thechance.mena.wallet.presentation.base.BaseViewModel
 import net.thechance.mena.wallet.presentation.base.ErrorState
 import net.thechance.mena.wallet.presentation.model.SubmissionStatus
@@ -16,7 +16,7 @@ import kotlin.uuid.Uuid
 
 @KoinViewModel
 class PaymentResultViewModel(
-    @Provided private val paymentRepository: PaymentRepository,
+    @Provided private val transactionRepository: TransactionRepository,
     @Provided private val paymentResultArgs: PaymentResultArgs,
     private val ioDispatcher: CoroutineDispatcher = Dispatchers.IO,
 ) : BaseViewModel<PaymentResultScreenState, PaymentResultEffect>(
@@ -56,7 +56,7 @@ class PaymentResultViewModel(
 
     private fun submitTransaction(transactionId: Uuid) {
         tryToExecute(
-            callee = { paymentRepository.submitTransaction(transactionId) },
+            callee = { transactionRepository.submitTransaction(transactionId) },
             onSuccess = { onSubmitTransactionSuccess() },
             onError = ::onSubmitTransactionFailed,
             dispatcher = ioDispatcher
